Kill ViewMixin
