+++ packages/@ember/application/instance.ts
@@ -8,10 +8,9 @@ import EngineInstance from '@ember/engine/instance';
 import type { BootOptions } from '@ember/engine/instance';
 import type Application from '@ember/application';
 import { renderSettled } from '@ember/-internals/glimmer';
-import type { BootEnvironment } from '@ember/-internals/glimmer';
+import type { BootEnvironment, Component } from '@ember/-internals/glimmer';
 import { assert } from '@ember/debug';
 import Router from '@ember/routing/router';
-import type { ViewMixin } from '@ember/-internals/views';
 import { EventDispatcher } from '@ember/-internals/views';
 import type { Registry } from '@ember/-internals/container';
 import type { SimpleElement } from '@simple-dom/interface';
@@ -149,7 +148,7 @@ class ApplicationInstance extends EngineInstance {
     @deprecated
     @private
   */
-  didCreateRootView(view: ViewMixin) {
+  didCreateRootView(view: Component) {
     view.appendTo(this.rootElement!);
   }
