+++ type-tests/@ember/component-test/component.ts
@@ -159,6 +159,6 @@ class SigExample extends Component<MySig> {
 // There is no type safety mapping `Element` here
 class ElementOnComponent extends Component<{ Element: HTMLDivElement }> {
   get hmm(): boolean {
-    return this.element.dispatchEvent(new Event('mousedown'));
+    return this.element?.dispatchEvent(new Event('mousedown')) ?? false;
   }
 }
